requirements.txt
streamlit
birdnetlib
google-generativeai
requests
tensorflow
resampy